Remove old antiquated tests, check for python 3, add tags.
